Add non Angular version of Events.
